+++ payments-core/src/main/java/com/stripe/android/view/CardFormView.kt
@@ -183,6 +183,23 @@ class CardFormView @JvmOverloads constructor(
     val paymentMethodCreateParams: PaymentMethodCreateParams?
         get() = paymentMethodCard?.let { PaymentMethodCreateParams.create(it) }
 
+    /**
+     * The Stripe account ID (if any) which is the business of record.
+     * See [use cases](https://docs.stripe.com/connect/charges#on_behalf_of) to determine if this option is relevant
+     * for your integration. This should match the
+     * [on_behalf_of](https://docs.stripe.com/api/payment_intents/create#create_payment_intent-on_behalf_of)
+     * provided on the Intent used when confirming payment.
+     */
+    var onBehalfOf: String? = null
+        set(value) {
+            if (isAttachedToWindow) {
+                doWithCardWidgetViewModel(viewModelStoreOwner) { viewModel ->
+                    viewModel.onBehalfOf = value
+                }
+            }
+            field = value
+        }
+
     init {
         orientation = VERTICAL